Parameter -> Variable, Act 2 -- this is probably the biggest change made to core BEAST in a few years. So I tagged revision 1870 for convenience. Parameter is now basically obsoleted and should be replaced by DoubleVariable or Variable<Double> where appropriate. We can now also develop some Variable<Integer> and Variable<Boolean> et cetera so should be nice in the future... I have also added a Variable<double[]> for vectors and matrices. See dr.inference.tutorial.Tutorial1 for a first peek at programmable BEAST MCMC...

+++ src/dr/inference/operators/UniformIntegerOperator.java
@@ -96,10 +96,10 @@ public class UniformIntegerOperator extends SimpleMCMCOperator {
 
             Parameter parameter = (Parameter) xo.getChild(Parameter.class);
 
-            int lower = (int) parameter.getBounds().getLowerLimit(0);
+            int lower = (int)(double)parameter.getBounds().getLowerLimit(0);
             if (xo.hasAttribute("lower")) lower = xo.getIntegerAttribute("lower");
 
-            int upper = (int) parameter.getBounds().getUpperLimit(0);
+            int upper = (int)(double)parameter.getBounds().getUpperLimit(0);
             if (xo.hasAttribute("upper")) upper = xo.getIntegerAttribute("upper");
 
             if (upper == lower || lower == (int) Double.NEGATIVE_INFINITY || upper == (int) Double.POSITIVE_INFINITY) {
